Added generic type. Wiped mocks until I'll figure them out.
